+++ src/huntsman/drp/dataid.py
@@ -1,16 +1,18 @@
 """ Classes to represent dataIds. """
+from collections import abc
 from contextlib import suppress
 
 from huntsman.drp.core import get_config
 
 
-class DataId(object):
+class DataId(abc.Mapping):
     """ A dataId behaves like a dictionary but makes it easier to compare between dataIds.
     DataId objects are hashable, whereas dictionaries are not. This allows them to be used in sets.
     """
-    _required_keys = None
+    _required_keys = tuple()
 
     def __init__(self, document, **kwargs):
+        super().__init__()
 
         # Check all the required information is present
         self._validate_document(document)
@@ -30,6 +32,18 @@ class DataId(object):
     def __getitem__(self, key):
         return self._document[key]
 
+    def __setitem__(self, key, item):
+        self._document[key] = item
+
+    def __delitem__(self, item):
+        del self._document[item]
+
+    def __iter__(self):
+        raise NotImplementedError
+
+    def __len__(self):
+        return len(self._document)
+
     def __str__(self):
         return str({k: self._document[k] for k in self._required_keys})
 
@@ -44,8 +58,8 @@ class DataId(object):
     def keys(self):
         return self.document.keys()
 
-    def to_dict(self):
-        return self._document.copy()
+    def update(self, d):
+        self._document.update(d)
 
     # Private methods
 
@@ -72,11 +86,9 @@ class RawExposureId(DataId):
 
 class CalibId(DataId):
 
-    _required_keys = ("calibDate", "datasetType", "filename")
+    _required_keys = ("calibDate", "datasetType", "filename", "ccd")
 
-    _required_keys_type = {"bias": ("calibDate", "ccd"),
-                           "dark": ("calibDate", "ccd"),
-                           "flat": ("calibDate", "ccd", "filter")}
+    _required_keys_type = {"flat": ("filter",)}
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
@@ -86,7 +98,9 @@ class CalibId(DataId):
         """
         super()._validate_document(document)
 
-        keys = self._required_keys_type[document["datasetType"]]
+        keys = self._required_keys_type.get(document["datasetType"], None)
+        if not keys:
+            return
 
         if not all([k in document for k in keys]):
             raise ValueError(f"Document does not contain all required keys: {keys}.")
